feat: add map tests
